mel.lib.common.TimeLogger: mv from mel.cmd.rotomapedit
